Renamed the folder and file naming to use cost_and_usage_report_*
